more int-type twiddling

+++ renderer.cpp
@@ -28,14 +28,14 @@ void Renderer::Render(const Parameters &parameters, const SpectrogramResults<dou
 	int numBins = spectrogramData.at(0).at(0).size();
 	int h = height - 1;
 	double colorScale = heatMapPalette.size() / -parameters.getDynRange();
-	size_t lastColorIndex = std::max(static_cast<size_t>(0), static_cast<size_t>(heatMapPalette.size()) - 1);
+	int lastColorIndex = std::max(0, static_cast<int>(heatMapPalette.size()) - 1);
 
 	for(int c = 0; c < 1 /*numChannels*/; c++) {
 		for(int x = 0; x < numSpectrums; x++) {
 			for(int y = 0; y < numBins; y++) {
 				int addr = x + (h - y) * stride32;
-				size_t colorindex = spectrogramData[c][x][y] * colorScale;
-				int32_t color = heatMapPalette[std::max(static_cast<size_t>(0), std::min(colorindex, lastColorIndex))];
+				int colorindex = spectrogramData[c][x][y] * colorScale;
+				int32_t color = heatMapPalette[std::max(0, std::min(colorindex, lastColorIndex))];
 				pixelBuffer[addr] = color;
 			}
 		}
